requirements.txt
pymongo==3.11.3
Flask_Cors==3.0.10
Flask==1.1.2
tensorflow-cpu
scikit_learn==0.24.2
gunicorn
dnspython